- Refactored gql types and resolvers into own folders
