+++ server/src/resolvers/index.ts
@@ -1,0 +1,3 @@
+import { userResolver } from './user'
+
+export const resolvers = [userResolver]
