+++ server/src/typeDefs/index.ts
@@ -1,0 +1,3 @@
+import { userType } from './user'
+
+export const typeDefs = [userType]
